+++ tasklist/webapp/src/main/java/io/camunda/tasklist/webapp/api/rest/v1/controllers/external/ProcessExternalController.java
@@ -20,8 +20,8 @@ import io.camunda.tasklist.webapp.rest.exception.Error;
 import io.camunda.tasklist.webapp.rest.exception.InvalidRequestException;
 import io.camunda.tasklist.webapp.rest.exception.NotFoundApiException;
 import io.camunda.tasklist.webapp.security.TasklistURIs;
-import io.camunda.tasklist.webapp.security.tenant.TenantService;
 import io.camunda.tasklist.webapp.service.ProcessService;
+import io.camunda.tasklist.webapp.tenant.TenantService;
 import io.camunda.webapps.schema.entities.ProcessEntity;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
