+++ tasklist/webapp/src/main/java/io/camunda/tasklist/webapp/security/identity/IdentityAuthentication.java
@@ -18,8 +18,8 @@ import io.camunda.tasklist.property.TasklistProperties;
 import io.camunda.tasklist.util.SpringContextHolder;
 import io.camunda.tasklist.webapp.security.OldUsernameAware;
 import io.camunda.tasklist.webapp.security.Permission;
-import io.camunda.tasklist.webapp.security.tenant.TasklistTenant;
-import io.camunda.tasklist.webapp.security.tenant.TenantAwareAuthentication;
+import io.camunda.tasklist.webapp.tenant.TasklistTenant;
+import io.camunda.tasklist.webapp.tenant.TenantAwareAuthentication;
 import java.util.Collections;
 import java.util.Date;
 import java.util.List;
