+++ tasklist/webapp/src/main/java/io/camunda/tasklist/webapp/security/oauth/IdentityTenantAwareJwtAuthenticationToken.java
@@ -10,8 +10,8 @@ package io.camunda.tasklist.webapp.security.oauth;
 import io.camunda.identity.sdk.Identity;
 import io.camunda.security.configuration.SecurityConfiguration;
 import io.camunda.tasklist.util.SpringContextHolder;
-import io.camunda.tasklist.webapp.security.tenant.TasklistTenant;
-import io.camunda.tasklist.webapp.security.tenant.TenantAwareAuthentication;
+import io.camunda.tasklist.webapp.tenant.TasklistTenant;
+import io.camunda.tasklist.webapp.tenant.TenantAwareAuthentication;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
