+++ tasklist/webapp/src/test/java/io/camunda/tasklist/webapp/api/rest/v1/controllers/external/ProcessExternalControllerTest.java
@@ -28,8 +28,8 @@ import io.camunda.tasklist.webapp.rest.exception.Error;
 import io.camunda.tasklist.webapp.rest.exception.InvalidRequestException;
 import io.camunda.tasklist.webapp.rest.exception.NotFoundApiException;
 import io.camunda.tasklist.webapp.security.TasklistURIs;
-import io.camunda.tasklist.webapp.security.tenant.TenantService;
 import io.camunda.tasklist.webapp.service.ProcessService;
+import io.camunda.tasklist.webapp.tenant.TenantService;
 import io.camunda.webapps.schema.entities.ProcessEntity;
 import io.camunda.webapps.schema.entities.form.FormEntity;
 import java.nio.charset.StandardCharsets;
